+++ spec/product.js
@@ -1,0 +1,41 @@
+describe('Occasion.Product', function() {
+  beforeEach(function() {
+    moxios.install();
+
+    window.onSuccess = jasmine.createSpy('onSuccess');
+    window.onFailure = jasmine.createSpy('onFailure');
+    window.onCompletion = jasmine.createSpy('onCompletion');
+
+    this.occsnClient = Occasion.Client({ token: 'my_token' });
+  });
+
+  afterEach(function() {
+    moxios.uninstall();
+  });
+
+  describe('attendeeQuestions', function() {
+    beforeEach(function () {
+      this.occsnClient.Product.find('1kbsdf')
+      .then(window.onSuccess);
+
+      this.promise = moxios.wait(() => {
+        return moxios.requests.mostRecent().respondWith(JsonApiResponses.Product.find.success)
+        .then(() => {
+          this.product = window.onSuccess.calls.mostRecent().args[0];
+        });
+      });
+    });
+
+    it('builds Collection of attendeeQuestions', function() {
+      return this.promise.then(() => {
+        expect(this.product.attendeeQuestions.empty()).toBeFalsy();
+      });
+    });
+
+    it('camelizes attendeeQuestions', function() {
+      return this.promise.then(() => {
+        expect(this.product.attendeeQuestions.first()).toEqual('firstName');
+      });
+    });
+  });
+});