+++ src/product.js
@@ -11,4 +11,10 @@ Occasion.Modules.push(function(library) {
   library.Product.hasMany('questions');
   library.Product.hasMany('redeemables');
   library.Product.hasMany('timeSlots');
+
+  library.Product.afterRequest(function() {
+    this.attendeeQuestions =
+      ActiveResource.Collection.build(this.attendeeQuestions)
+      .map((q) => { return s.camelize(q, true); });
+  });
 });